stage4_mcp_agent/requirements.txt
langchain 
langgraph 
langchain-openai 
python-dotenv
mcp